QualificationsRepository added. Moved Qualifications methods into Repository. Deleted ContentfulFilterService and methods on ContentfulContentService. Shared functionality moved into base service class.
