--- src/Dfe.EarlyYearsQualification.Content/Services/IContentFilterService.cs
@@ -1,8 +1,0 @@
-using Dfe.EarlyYearsQualification.Content.Entities;
-
-namespace Dfe.EarlyYearsQualification.Content.Services;
-
-public interface IContentFilterService
-{
-    Task<List<Qualification>> GetFilteredQualifications(int? level, int? startDateMonth, int? startDateYear, string? awardingOrganisation, string? qualificationName);
-}
