+++ src/Dfe.EarlyYearsQualification.Web/Controllers/ConfirmQualificationController.cs
@@ -14,6 +14,7 @@ namespace Dfe.EarlyYearsQualification.Web.Controllers;
 [RedirectIfDateMissing]
 public class ConfirmQualificationController(
     ILogger<ConfirmQualificationController> logger,
+    IQualificationsRepository qualificationsRepository,
     IContentService contentService,
     IUserJourneyCookieService userJourneyCookieService,
     IGovUkContentParser contentParser)
@@ -21,10 +22,10 @@ public class ConfirmQualificationController(
 {
     [HttpGet]
     [Route("{qualificationId}")]
-    #pragma warning disable S6967
+#pragma warning disable S6967
     // ...the model is a string, so no need to check ModelState.IsValid here
     public async Task<IActionResult> Index(string qualificationId)
-    #pragma warning restore S6967
+#pragma warning restore S6967
     {
         if (string.IsNullOrEmpty(qualificationId))
         {
@@ -39,7 +40,7 @@ public class ConfirmQualificationController(
             return RedirectToAction("Index", "Error");
         }
 
-        var qualification = await contentService.GetQualificationById(qualificationId);
+        var qualification = await qualificationsRepository.GetQualificationById(qualificationId);
         if (qualification is null)
         {
             var loggedQualificationId = qualificationId.Replace(Environment.NewLine, "");
@@ -63,7 +64,7 @@ public class ConfirmQualificationController(
             return RedirectToAction("Index", "Error");
         }
 
-        var qualification = await contentService.GetQualificationById(model.QualificationId);
+        var qualification = await qualificationsRepository.GetQualificationById(model.QualificationId);
         if (qualification is null)
         {
             var loggedQualificationId = model.QualificationId.Replace(Environment.NewLine, "");
@@ -135,7 +136,8 @@ public class ConfirmQualificationController(
                    QualificationDateAdded = qualification.FromWhichYear!,
                    BackButton = MapToNavigationLinkModel(content.BackButton),
                    PostHeadingContent = await contentParser.ToHtml(content.PostHeadingContent),
-                   VariousAwardingOrganisationsExplanation = await contentParser.ToHtml(content.VariousAwardingOrganisationsExplanation)
+                   VariousAwardingOrganisationsExplanation =
+                       await contentParser.ToHtml(content.VariousAwardingOrganisationsExplanation)
                };
     }
 }
