+++ tests/Dfe.EarlyYearsQualification.UnitTests/Controllers/QuestionsControllerTests.cs
@@ -28,16 +28,16 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = controller.StartNew();
-        
+
         result.Should().NotBeNull();
 
         var resultType = result as RedirectToActionResult;
@@ -47,7 +47,7 @@ public class QuestionsControllerTests
 
         mockUserJourneyCookieService.Verify(x => x.ResetUserJourneyCookie(), Times.Once);
     }
-    
+
     [TestMethod]
     public async Task WhereWasTheQualificationAwarded_ContentServiceReturnsNoQuestionPage_RedirectsToErrorPage()
     {
@@ -55,7 +55,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -63,7 +63,7 @@ public class QuestionsControllerTests
                           .ReturnsAsync((RadioQuestionPage?)default).Verifiable();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhereWasTheQualificationAwarded();
@@ -88,7 +88,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -102,7 +102,7 @@ public class QuestionsControllerTests
                           .ReturnsAsync(questionPage);
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhereWasTheQualificationAwarded();
@@ -132,7 +132,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -140,7 +140,7 @@ public class QuestionsControllerTests
                           .ReturnsAsync(new RadioQuestionPage());
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         controller.ModelState.AddModelError("option", "test error");
@@ -168,12 +168,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result =
@@ -191,8 +191,9 @@ public class QuestionsControllerTests
         resultType!.ActionName.Should().Be("QualificationOutsideTheUnitedKingdom");
         resultType.ControllerName.Should().Be("Advice");
 
-        mockUserJourneyCookieService.Verify(x => x.SetWhereWasQualificationAwarded(QualificationAwardLocation.OutsideOfTheUnitedKingdom),
-                                            Times.Once);
+        mockUserJourneyCookieService
+            .Verify(x => x.SetWhereWasQualificationAwarded(QualificationAwardLocation.OutsideOfTheUnitedKingdom),
+                    Times.Once);
     }
 
     [TestMethod]
@@ -202,12 +203,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result =
@@ -233,12 +234,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result =
@@ -264,12 +265,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result =
@@ -284,8 +285,9 @@ public class QuestionsControllerTests
         resultType!.ActionName.Should().Be("QualificationsAchievedInNorthernIreland");
         resultType.ControllerName.Should().Be("Advice");
 
-        mockUserJourneyCookieService.Verify(x => x.SetWhereWasQualificationAwarded(QualificationAwardLocation.NorthernIreland),
-                                            Times.Once);
+        mockUserJourneyCookieService
+            .Verify(x => x.SetWhereWasQualificationAwarded(QualificationAwardLocation.NorthernIreland),
+                    Times.Once);
     }
 
     [TestMethod]
@@ -295,12 +297,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result =
@@ -325,7 +327,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -344,7 +346,7 @@ public class QuestionsControllerTests
         mockPlaceholderUpdater.Setup(x => x.Replace(It.IsAny<string>())).Returns<string>(x => x);
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhenWasTheQualificationStarted();
@@ -373,7 +375,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -381,7 +383,7 @@ public class QuestionsControllerTests
                           .ReturnsAsync((DateQuestionPage?)default).Verifiable();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhenWasTheQualificationStarted();
@@ -403,7 +405,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -420,7 +422,7 @@ public class QuestionsControllerTests
                           .ReturnsAsync(questionPage);
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         mockQuestionModelValidator.Setup(x => x.IsValid(It.IsAny<DateQuestionModel>(), It.IsAny<DateQuestionPage>()))
@@ -459,7 +461,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -468,7 +470,7 @@ public class QuestionsControllerTests
             .Returns(new ValidationResult { IsValid = true });
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         const int selectedMonth = 12;
@@ -499,7 +501,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -507,7 +509,7 @@ public class QuestionsControllerTests
                           .ReturnsAsync((RadioQuestionPage?)default).Verifiable();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatLevelIsTheQualification();
@@ -532,7 +534,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -540,7 +542,11 @@ public class QuestionsControllerTests
                            {
                                Question = "Test question",
                                CtaButtonText = "Continue",
-                               Options = [new Option { Label = "Label", Value = "Value" }, new Divider { Text = "Test" }],
+                               Options =
+                               [
+                                   new Option { Label = "Label", Value = "Value" },
+                                   new Divider { Text = "Test" }
+                               ],
                                AdditionalInformationHeader = "Test header",
                                AdditionalInformationBody = ContentfulContentHelper.Text("Test html body")
                            };
@@ -550,7 +556,7 @@ public class QuestionsControllerTests
         mockContentParser.Setup(x => x.ToHtml(It.IsAny<Document>())).ReturnsAsync("Test html body");
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatLevelIsTheQualification();
@@ -587,15 +593,15 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         mockContentService.Setup(x => x.GetRadioQuestionPage(QuestionPages.WhatLevelIsTheQualification))
                           .ReturnsAsync(new RadioQuestionPage());
-        
+
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         controller.ModelState.AddModelError("option", "test error");
@@ -609,7 +615,7 @@ public class QuestionsControllerTests
         resultType!.ViewName.Should().Be("Radio");
 
         var model = resultType.Model as RadioQuestionModel;
-        
+
         model.Should().NotBeNull();
         model!.HasErrors.Should().BeTrue();
 
@@ -623,12 +629,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatLevelIsTheQualification(new RadioQuestionModel
@@ -653,7 +659,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -661,7 +667,7 @@ public class QuestionsControllerTests
                                     .Returns(true);
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatLevelIsTheQualification(new RadioQuestionModel
@@ -685,7 +691,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -693,7 +699,7 @@ public class QuestionsControllerTests
                                     .Returns(true);
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatLevelIsTheQualification(new RadioQuestionModel
@@ -717,7 +723,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -725,7 +731,7 @@ public class QuestionsControllerTests
                                     .Returns(false);
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatLevelIsTheQualification(new RadioQuestionModel
@@ -749,12 +755,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatLevelIsTheQualification(new RadioQuestionModel
@@ -778,7 +784,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -786,7 +792,7 @@ public class QuestionsControllerTests
                           .ReturnsAsync((DropdownQuestionPage?)default).Verifiable();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatIsTheAwardingOrganisation();
@@ -811,7 +817,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -828,7 +834,7 @@ public class QuestionsControllerTests
         mockContentService.Setup(x => x.GetDropdownQuestionPage(QuestionPages.WhatIsTheAwardingOrganisation))
                           .ReturnsAsync(questionPage);
 
-        mockContentFilterService
+        mockRepository
             .Setup(x => x.GetFilteredQualifications(
                                                     It.IsAny<int?>(),
                                                     It.IsAny<int?>(),
@@ -838,7 +844,7 @@ public class QuestionsControllerTests
             .ReturnsAsync([]);
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatIsTheAwardingOrganisation();
@@ -870,7 +876,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -896,13 +902,13 @@ public class QuestionsControllerTests
                                            "B awarding organisation", 123)
                                    };
 
-        mockContentFilterService
+        mockRepository
             .Setup(x => x.GetFilteredQualifications(It.IsAny<int?>(), It.IsAny<int?>(), It.IsAny<int?>(),
                                                     It.IsAny<string?>(), It.IsAny<string?>()))
             .ReturnsAsync(listOfQualifications);
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatIsTheAwardingOrganisation();
@@ -936,7 +942,7 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
@@ -960,13 +966,13 @@ public class QuestionsControllerTests
                                            AwardingOrganisations.AllHigherEducation, 123)
                                    };
 
-        mockContentFilterService
+        mockRepository
             .Setup(x => x.GetFilteredQualifications(It.IsAny<int?>(), It.IsAny<int?>(), It.IsAny<int?>(),
                                                     It.IsAny<string?>(), It.IsAny<string?>()))
             .ReturnsAsync(listOfQualifications);
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatIsTheAwardingOrganisation();
@@ -993,12 +999,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var questionPage = new DropdownQuestionPage
@@ -1014,7 +1020,7 @@ public class QuestionsControllerTests
         mockContentService.Setup(x => x.GetDropdownQuestionPage(QuestionPages.WhatIsTheAwardingOrganisation))
                           .ReturnsAsync(questionPage);
 
-        mockContentFilterService
+        mockRepository
             .Setup(x => x.GetFilteredQualifications(It.IsAny<int?>(), It.IsAny<int?>(), It.IsAny<int?>(),
                                                     It.IsAny<string?>(), It.IsAny<string?>()))
             .ReturnsAsync([]);
@@ -1044,12 +1050,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var result = await controller.WhatIsTheAwardingOrganisation(new DropdownQuestionModel
@@ -1076,12 +1082,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var questionPage = new DropdownQuestionPage
@@ -1096,8 +1102,6 @@ public class QuestionsControllerTests
 
         mockContentService.Setup(x => x.GetDropdownQuestionPage(QuestionPages.WhatIsTheAwardingOrganisation))
                           .ReturnsAsync(questionPage);
-
-        mockContentService.Setup(x => x.GetQualifications()).ReturnsAsync([]);
 
         var result = await controller.WhatIsTheAwardingOrganisation(new DropdownQuestionModel
                                                                     {
@@ -1126,12 +1130,12 @@ public class QuestionsControllerTests
         var mockContentService = new Mock<IContentService>();
         var mockContentParser = new Mock<IGovUkContentParser>();
         var mockUserJourneyCookieService = new Mock<IUserJourneyCookieService>();
-        var mockContentFilterService = new Mock<IContentFilterService>();
+        var mockRepository = new Mock<IQualificationsRepository>();
         var mockQuestionModelValidator = new Mock<IDateQuestionModelValidator>();
         var mockPlaceholderUpdater = new Mock<IPlaceholderUpdater>();
 
         var controller = new QuestionsController(mockLogger.Object, mockContentService.Object, mockContentParser.Object,
-                                                 mockUserJourneyCookieService.Object, mockContentFilterService.Object,
+                                                 mockUserJourneyCookieService.Object, mockRepository.Object,
                                                  mockQuestionModelValidator.Object, mockPlaceholderUpdater.Object);
 
         var questionPage = new DropdownQuestionPage
@@ -1146,8 +1150,6 @@ public class QuestionsControllerTests
 
         mockContentService.Setup(x => x.GetDropdownQuestionPage(QuestionPages.WhatIsTheAwardingOrganisation))
                           .ReturnsAsync(questionPage);
-
-        mockContentService.Setup(x => x.GetQualifications()).ReturnsAsync([]);
 
         var result = await controller.WhatIsTheAwardingOrganisation(new DropdownQuestionModel
                                                                     {
